+++ Clients/OIViewer/main.cpp
@@ -44,7 +44,7 @@ int mainFunction(int argc, const wchar_t** argv)
             filePath = LLUtils::FileSystemHelper::ResolveFullPath(filePath);
             COPYDATASTRUCT copyData{};
             copyData.dwData = ::OIV::Win32::UserMessage::PRIVATE_WM_LOAD_FILE_EXTERNALLY;
-            copyData.cbData = (filePath.length() + 1) * sizeof(decltype(filePath)::value_type);
+            copyData.cbData = static_cast<DWORD>((filePath.length() + 1) * sizeof(decltype(filePath)::value_type));
             copyData.lpData = const_cast<wchar_t*>(filePath.c_str());
             ::SendMessage(window, WM_COPYDATA, ::OIV::Win32::UserMessage::PRIVATE_WM_LOAD_FILE_EXTERNALLY, (LPARAM)(LPVOID)&copyData);
         }
